refactor: split NginxConfigRenderer out of managers/nginx.py
nginx.py mixed config-text rendering with install/reload side effects
in one 892-line file. Renderer now lives in nginx_render.py; nginx.py
re-exports it so no import site elsewhere needs to change.
